Add base configuration of ocelot
